Refactor Set._tostring. Fixes #61
__tostring convertation metatable function from the code imported from
lua-set library has been reworked by the code, based on the idea,
of the one, that was sugested by @GitSparTV in #61.

The purpose of that refactoring is to reduce memory usage,
caused by strings reallocation by string concatenation.
(Table concatenation cheaper)

Despite code should be pretty valid, and passes all tests,
I think it needs to survive a bunch of real-life tests,
but, unlikely, I have not enough spare time ATM.

Please, test it and report, if any.

Signed-off-by: Vadim A. Misbakh-Soloviov <git@mva.name>